chore: move App component from components/ to pages/
