+++ src/pages/App/index.test.js
